Move degree and timeline types to shared
- move Back-End/src/types/timeline.ts → shared/types/timeline.ts.
- Add shared/types/degree.ts
