+++ Back-End/src/services/audit/degreeAuditService.ts
@@ -1,9 +1,5 @@
 import { Timeline, User } from '@models';
-import {
-  degreeController,
-  CoursePoolInfo,
-  DegreeData,
-} from '@controllers/degreeController';
+import { degreeController } from '@controllers/degreeController';
 import { CourseData } from '@controllers/courseController';
 import { CourseStatus } from '../../types/transcript';
 import {
@@ -17,7 +13,8 @@ import {
   RequirementStatus,
   GenerateAuditParams,
 } from '@shared/audit';
-import { TimelineCourse, TimelineDocument, TimelineResult, TimelineSemester } from '../../types/timeline';
+import { TimelineCourse, TimelineDocument, TimelineResult, TimelineSemester } from '@shared/timeline';
+import {DegreeData, CoursePoolInfo} from '@shared/degree';
 import { isTermInProgress } from '@utils/misc';
 
 interface TimelineWithUser extends TimelineDocument {
