+++ Back-End/src/utils/pythonUtilsApi.ts
@@ -1,6 +1,6 @@
 import { CourseData } from '@controllers/courseController';
 import { CoursePoolData } from '@controllers/coursepoolController';
-import { DegreeData } from '@controllers/degreeController';
+import { DegreeData } from '@shared/degree';
 import { ParsedData } from '../types/transcript';
 import { PYTHON_SERVICE_BASE_URL } from '@utils/constants';
 import axios from 'axios';
