+++ shared/types/degree.ts
@@ -1,0 +1,13 @@
+
+export interface DegreeData {
+  _id: string;
+  name: string;
+  totalCredits: number;
+  coursePools?: string[];
+}
+export interface CoursePoolInfo {
+  _id: string;
+  name: string;
+  creditsRequired: number;
+  courses: string[];
+}
